Insert Navbar in search page
