+++ src/app/search/page.tsx
@@ -1,64 +1,22 @@
-"use client";
-import { Suspense } from 'react';
-import { useSearchParams } from 'next/navigation';
-import DashboardPage from '../dashboard/page';
-import Navbar from '@/components/user/perfil/navbar/Navbar';
-import SearchBar from '@/components/user/perfil/navbar/SearchBar';
-import styles from '@/app/dashboard/dashboard.module.css';
-import useSWR from 'swr';
-import UsersList from '@/components/dashboard/UsersList';
-import Link from 'next/link';
+import { useEffect, useState } from 'react';
+import { NextPageContext } from "next";
+import styles from "./search.module.css";
+import getCurrentUser from "@/app/actions/getCurrentUser";
+import Navbar from "@/components/user/perfil/navbar/Navbar";
+import SearchResults from '@/components/user/perfil/navbar/SearchResults';
 
-const fetchPosts = async (url: string) => {
-    const response = await fetch(url);
 
-    if (!response.ok) {
-        throw new Error('Failed to fetch posts');
-    }
-
-    return response.json();
-}
-
-const SearchPage = () => {
-  const search = useSearchParams();
-  const searchQuery = search ? search?.get('q') : null;
-  const encodedSearchQuery = encodeURI(searchQuery || '');
-  
-  const { data, isLoading } = useSWR(`/api/search?q=${encodedSearchQuery}`, fetchPosts);
-
-  
-
-  console.log('La data es:', data);
+async function SearchPage() {
+  const currentUser = await getCurrentUser();
 
   return (
-    <>
-    <div className={styles.bgPrattle}>
-    <nav className="flex w-full items-center p-4">
-        <div className="my-2 mx-3 hidden lg:block">
-          <Link href="/dashboard">
-            <img src="/img/logo_blanc.png" className="h-24 w-auto object-contain cursor-pointer" alt="Logo" />
-          </Link>
+      <div className={`h-screen ${styles.bgPrattle}`}>
+        <Navbar currentUser={currentUser!} />
+        <div>
+          <SearchResults />
         </div>
-        <SearchBar />
-    </nav>
-    </div>
-    <main className={`h-screen ${styles.bgPrattle}`}>
-        {isLoading && <p>Loading...</p>}
-        {!isLoading && data && data.users && (
-          <div>
-            <UsersList title="Resultados de busqueda" items={data.users}/>
-          </div>
-        )}
-        {!isLoading && (!data || !data.users) && <p>No results found.</p>}
-      </main>
-    </>
+      </div>
   );
-};
+}
 
-const SuspenseWrapper = () => (
-  <Suspense fallback={<div>Loading...</div>}>
-    <SearchPage />
-  </Suspense>
-);
-
-export default SuspenseWrapper;
+export default SearchPage;
