+++ src/app/search/search.module.css
@@ -1,0 +1,6 @@
+.bgPrattleclear {
+    background-color: #29313df8;
+}
+.bgPrattle {
+    background-color: #1f2836;
+  }
